backend/similarity-fashion-search/requirements.txt
weaviate-client==3.9.0

Flask==2.2.2

Pillow~=8.2.0
tensorflow~=2.8.0
numpy~=1.20.1